Moved the test into test/ to keep *.h consistent.
