+++ cerealize/test/test_event_base.h
@@ -7,7 +7,7 @@
 #include <sstream>
 #include <memory>
 
-#include "cerealize.h"
+#include "../cerealize.h"
 
 struct MapsYouEventBase {
   std::string uid = "";
@@ -32,7 +32,8 @@ struct MapsYouEventBase {
   void serialize(A& ar) {
     // `NVP` below stands for Name-Value-Pair.
     // The `CEREAL_NVP` syntax is only to have JSON-based serialization contain human-readable field names.
-    // The binary format is identical, no matter whether `uid`, `CEREAL_NVP(uid)` or `cereal::make_nvp("foo", uid)`.
+    // The binary format is identical, no matter whether `uid`, `CEREAL_NVP(uid)` or `cereal::make_nvp("foo",
+    // uid)`.
     // For binary serialization format, the order and types of the fields should stay the same.
     ar(CEREAL_NVP(uid), CEREAL_NVP(uid_facebook), CEREAL_NVP(uid_google), CEREAL_NVP(uid_apple));
   }
